requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0